refactor(markdown-navigator): rename files
