+++ src/app/components/components/markdown-navigator/markdown-navigator.component.scss
