+++ src/platform/markdown-navigator/markdown-navigator-window/markdown-navigator-window.component.scss
