+++ src/platform/markdown-navigator/markdown-navigator.component.scss
