+++ src/platform/markdown-navigator/public-api.ts
@@ -1,4 +1,4 @@
-export * from './md-navigator.module';
-export * from './md-navigator.component';
-export * from './md-navigator-window/md-navigator-window.component';
-export * from './md-navigator-window-service/md-navigator-window.service';
+export * from './markdown-navigator.module';
+export * from './markdown-navigator.component';
+export * from './markdown-navigator-window/markdown-navigator-window.component';
+export * from './markdown-navigator-window-service/markdown-navigator-window.service';
